backend/requirements.txt
asgiref==3.7.2
certifi==2024.7.4
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.3
# django-ckeditor==6.7.1
# django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.14.0
django_ckeditor_5
django-ckeditor
filelock==3.15.4
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
ply==3.11
protobuf==3.20.3
pytz==2024.1
setuptools==69.5.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3
wheel==0.43.0
whitenoise==6.6.0
